Allow routes to render to be specified in `routeKey` query string parameters (#10)
Note that this also only renders links to the specified routes, whereas
we probably want to render all of them, or perhaps some hardcoded list.

---

This allows the original URL format to work, so that e.g. http://127.0.0.1:61015/manhattan-country-school-too shows the `manhattan-country-school-too` route, and so that the iframe in the beacon page at e.g. http://127.0.0.1:61015/beacon/manhattan-country-school-too/D3ADB33F continues to work as before.

One janky thing about this setup is that if `routeKey` is specified as part of the URL query, then it will take precedence over the value in the URL path. For example, http://127.0.0.1:61015/manhattan-country-school?routeKey=manhattan-country-school-too will render only the `manhattan-country-school-too` route, and ignore the `manhattan-country-school` in the URL entirely.

It would probably be ideal if the beacon page's iframe specified a URL with the `routeKey` in the URL query, rather than directly in the path, so that we can avoid specifying the routeKey in the URL path at all, but I haven't gone that far yet. Assuming things are now working as expected in the manual test cases, I think it might be a good time to figure out how to write automated tests for them, then continue refactoring.

---

## Instructions to Setup a Testing Environment ##

I've invoked the server like this while testing:
`PORT=61015 beacon_hash=D3ADB33F yarn start`

And then I discovered that Chrome lets you mock geo-locations out of the box & w/o any extensions (thanks for nothing, Safari). #bless this kind stranger for [her Dev•to turorial](https://dev.to/domizajac/how-to-mock-location-in-the-web-without-installing-additional-tools-50j8). tl;dr:

1. Open Chrome
2. Go to page you want to test
3. Open Chrome DevTools (`ctrl + shift + I` on Linux/Windows or `⌘ + shift + I` on Mac)
4. Click 3 dots on the right side.
5. Choose `More Tools` and `Sensors`
6. Check `Geolocation` section

Now open a private window/separate session to start the beacon broadcast by going to this URL: 
http://127.0.0.1:61015/beacon/manhattan-country-school/D3ADB33F

Useful LAT/LONGs for testing this route:
* Marcus Garvey Park (SW corner): `40.803917, -73.946054` 
* NW corner of Central Park: `40.799781, -73.957090`
* Columbus Ave @ W 85th: `40.785835, -73.972702`

And another private window/separate session for the `manhattan-country-school-too` beacon: 
http://127.0.0.1:61015/beacon/manhattan-country-school-too/D3ADB33F

Useful LAT/LONGs for testing this route:
* Marcus Garvey Park (SW corner): `40.803917, -73.946054` 
* NE corner of Central Park: `40.796907, -73.949352`
* 5th Ave @ E 86th: `40.781518, -73.960395`

Now open the combine route page at: 
http://127.0.0.1:61015/?routeKey=manhattan-country-school&routeKey=manhattan-country-school-too

---

In the future, we might want to change the color of the bus position markers to match the color of the route's path.

---

I actually joined the bike bus this morning! Ryan Christman was there and we got coffee afterwards, and I told him about the changes we're working on here.

While discussing this marker color thing, I realized that it might not be all that important that the markers be the same color as the route, since presumably they will be very close to each other until the routes split, so it doesn't really matter which marker belongs to which route, and once the routes split, it will be clear which marker belongs to each route.

That said, I did some research into how to change the color, and it looks like we could use Leaflet's `divIcon` method to create a custom `<div>` containing an `<img>` of the marker icon (see https://stackoverflow.com/a/53637331), then we could use a CSS `filter` to change the color of the icon to match the color of the route (see https://stackoverflow.com/a/67615509). Or, we could use SVG directly to draw an icon with the #RRGGBB color of the route, without needing the `<img>` and CSS `filter`.

---

Co-authored-by: Dustø Masterson Goldman <dustin@goldmans.net>

---------

Co-authored-by: Dustin Goldman <dustin@goldmans.net>