filter published routes at start up
instead of handling the filtering logic for "published" routes in the `index.hbs` Handle Template, we now filter the routes once at start up, save those routes as a `const`, and pass that filtered object onto to `index.hbs`. Then in `index.hbs`, we render _all_ route objects found in the `routes` object.