add `combinedRouteKeys` support to tracker
add support for a "meta" route, which has meta data for a route, but does not have a `stops` array. Instead it has property named `combinedRouteKeys`, which is an array of strings that are keys for actual routes.

When rendering the `tracker` page, the "meta" route is passed as the `route` value, and all the routes referenced by `combinedRouteKeys` are passed as the `routes` value.