Remove `node-persist`, just use `nano-cache`
Storing the coordinates to disk seems unnecessary, especially since they
won't persist across deploys or other Heroku restarts. This will also
make it easier to have the coordinates automatically expire after a
certain amount of time, as `nano-cache` has built-in support for that,
and we won't need to worry about what to do with the `node-persist`ed
copies of the coords.