Clear markers after 30m (#36)
* Remove `node-persist`, just use `nano-cache`
* Add failing test for expiring coordinates
* yarn add --dev @sinonjs/fake-timers
* Change TTL to 30m
* Fake time passage to fix expiry test

Storing the coordinates to disk seems unnecessary, especially since they
won't persist across deploys or other Heroku restarts. This will also
make it easier to have the coordinates automatically expire after a
certain amount of time, as `nano-cache` has built-in support for that,
and we won't need to worry about what to do with the `node-persist`ed
copies of the coords.

---------

Co-authored-by: Dustø Masterson Goldman <dustin@goldmans.net>